move projection class out of ce package
