+++ app/server/appsmith-server/src/test/java/com/appsmith/server/repositories/CustomUserDataRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.appsmith.server.repositories;
 
 import com.appsmith.server.domains.UserData;
-import com.appsmith.server.projections.ce.UserDataProfilePhotoProjection;
+import com.appsmith.server.projections.UserDataProfilePhotoProjection;
 import com.mongodb.client.result.UpdateResult;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
